ledger/blockstore: Use `u32` for shred indices.
Shred indices, within a slot, can not exceed a `u32` value.
Unfortunately, at the moment, there is no common type that is used when
a shred index needs to be stored.  Part of the code use `u32`, another
considerable part uses `u64` and a few places use `usize`.

Ideally, it would be better to have a dedicated type, wrapping the
primitive `u32` value.

This change is trying to unify the types used, moving most of the code
that works with shred indices to `u32`.  `u64` is used in a few
locations that are serialized into the blockstore or are part of the
wire protocol.  Those are left as is for now.

As the conversion is not 100%, there are still a few conversions left,
that could be avoided should both the serialization and the wire
protocol code use `u32`.  I also was not completely sure about the
`usize` types used by the Merkel tree code - those might be convertible
to `u32` as well.

Stats (roughly):

  * Removed `as u{32,64}`:

  $ git diff --cached | grep '^-.*as u\(32\|64\)' | wc -l
  50

  * Removed `unwrap()`/`expect()`:

  $ git diff --cached | grep '^-.*\(unwrap()\|expect()\)' | wc -l
  20

  * Added `as u{32,64}`:

  $ git diff --cached | grep '^+.*as u\(32\|64\)' | wc -l
  7

  * Added `unwrap()`/`expect()`:

  $ git diff --cached | grep '^+.*\(unwrap()\|expect()\)' | wc -l
  32

Total is 50 + 20 - 8 - 32 = 30 conversions removed.  And a lot of the
removed conversion are unchecked, while a lot of the added ones are
checked.  Current version would still panic, if a value outside the
`u32` scope is used - it will just panic further away from the bug
location, complicating the debugging efforts.

This change also prepares the grounds for removal of most of the
conversion that are still present.  We would need to convert the
serialization and the wire code to `u32` as well.

There are a number of `as usize` conversions added in the new code.
Those should be safe, as they convert from `u32`.  Most of them are used
when indexing into `Vec`s.

+++ ledger/src/blockstore_processor.rs
@@ -1268,7 +1268,7 @@ impl ReplaySlotStats {
         slot: Slot,
         num_txs: usize,
         num_entries: usize,
-        num_shreds: u64,
+        num_shreds: u32,
         bank_complete_time_us: u64,
         is_unified_scheduler_enabled: bool,
     ) {
@@ -1389,7 +1389,7 @@ impl ReplaySlotStats {
 pub struct ConfirmationProgress {
     pub last_entry: Hash,
     pub tick_hash_count: u64,
-    pub num_shreds: u64,
+    pub num_shreds: u32,
     pub num_entries: usize,
     pub num_txs: usize,
 }
@@ -1455,7 +1455,7 @@ pub fn confirm_slot(
 fn confirm_slot_entries(
     bank: &BankWithScheduler,
     replay_tx_thread_pool: &ThreadPool,
-    slot_entries_load_result: (Vec<Entry>, u64, bool),
+    slot_entries_load_result: (Vec<Entry>, u32, bool),
     timing: &mut ConfirmationTiming,
     progress: &mut ConfirmationProgress,
     skip_verification: bool,
